Segregate generics into another interface
