+++ api/src/test/java/io/victoralbertos/jolyglot/Json.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2016 Victor Albertos
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.victoralbertos.jolyglot;
+
+
+class Json {
+
+
+  String jsonMockParameterizedSample() {
+    return "{\"t\":{\"s1\":\"s1\"},\"s1\":\"s1\"}";
+  }
+
+  String jsonMockParameterizedSampleReverse() {
+    return "{\"s1\":\"s1\",\"t\":{\"s1\":\"s1\"}}";
+  }
+
+}
